map H and L
Map H in normal mode to go to the beginning of the current line. Since h moves left you can think of H as a "stronger" h.
Map L in normal mode to go to the end of the current line. Since l moves right you can think of L as a "stronger" l.